Add files via upload
These archives contain the lists corresponding to the semantic expansion expressed in the file name of the Yelp taxonomy category labels.

In example. autogeneratedLists_SYN.zip contains as many lists as the Yelp taxonomy category labels cardinality, each named as a category label and each contained in a folder named as the macro-category of the category label. Each list contains the label of the category, the label of the macrocategory and all the synonyms of them, plus the lemmas of them, normalized by number and weighted.